refactor(subway): 도메인, 레포지토리 별로 패키지 분리
