+++ src/main/java/subway/domain/line/Line.java
@@ -1,9 +1,9 @@
-package subway.domain;
+package subway.domain.line;
 
-public class Station {
+public class Line {
     private String name;
 
-    public Station(String name) {
+    public Line(String name) {
         this.name = name;
     }
