+++ src/main/java/subway/repository/line/LineRepository.java
@@ -1,4 +1,6 @@
-package subway.domain;
+package subway.repository.line;
+
+import subway.domain.line.Line;
 
 import java.util.ArrayList;
 import java.util.Collections;
